+++ src/js/constants/LocalVolumesConstants.js
@@ -1,0 +1,12 @@
+const LocalVolumesConstants = {
+  TYPES: {
+    DOCKER: "DOCKER",
+    LOCAL: "LOCAL"
+  },
+  STATUS: {
+    ATTACHED: "Attached",
+    DETACHED: "Detached"
+  }
+};
+
+export default Object.freeze(LocalVolumesConstants);