+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - \#3157 - Filter apps by attached volumes
 - \#3127 - Show groups as part of search results
 - \#966  - Allow creating empty groups
+- \#3240 - Add volumes tab to application page
 
 ### Changed
 - \#3078 - Always show download button for logs